+++ packages/terra-date-picker/CHANGELOG.md
@@ -8,6 +8,7 @@ Unreleased
 ------------------
 ### Changed
 * Removed use of `componentWillReceiveProps`.
+* Added ability to shift focus around interactive elements within the datepicker popup to enable better screen reader/keyboard usage
 * Removed `details` tag from doc-site.
 * Doc site example module import syntax modified to use package.json name resolution